requirements.txt
python-telegram-bot[job-queue]>=22.0
aiosqlite>=0.19.0
aiohttp>=3.10.0
feedparser>=6.0.11
trafilatura>=1.12.2
beautifulsoup4>=4.12.3
python-dateutil>=2.9.0
rapidfuzz>=3.9.7
python-dotenv>=1.0.1
groq>=0.18.0
pytest>=8.3.2
pytest-asyncio>=0.24.0

